requirements.txt
alfred-workflow>=1.40.0 